✨Feat: restaurants auto complete (#5)
* ✨  feat: statistique user

* 🚚 fix: rename UserRestaurantStatisticsDto

* ⚡ fix: appel Firebase depuis le back

* ⚡ fix: simplification des méthodes

* :zap: fix: top 3 en requêtes SQL

* 🚚 fix: suppression DTO inutile

* ♻️ fix: refacto connexion (signIn/up dans le front via firebase)

* ✨ feat: get all restaurant

+++ src/main/java/com/forkify_backend/api/dto/UserStatisticsDto.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 @Setter
@@ -14,7 +16,7 @@ import lombok.Setter;
 public class UserStatisticsDto {
     @NotNull
     @PositiveOrZero
-    Double amountSpent;
+    BigDecimal amountSpent;
 
     @NotNull
     @PositiveOrZero
